@@ -6,9 +6,9 @@
   "dependencies": {
     "@mdi/js": "^3.3.92",
     "gh-pages": "^2.0.1",
-    "react": "^16.7.0",
-    "react-dom": "^16.7.0",
-    "react-scripts": "2.1.3"
+    "react": "^16.13.0",
+    "react-dom": "^16.13.0",
+    "react-scripts": "3.4.0"
   },
   "scripts": {
     "predeploy": "npm run build",
